refactor: create a common function for git authentication
Update the Create Unlocked and Source Package tasks to use a common function for
authentication to version control provider.
