+++ packages/azpipelines/BuildTasks/CreateUnlockedPackageTask/CreateUnlockedPackage.ts
@@ -1,6 +1,7 @@
 import tl = require("azure-pipelines-task-lib/task");
 import CreateUnlockedPackageImpl from "@dxatscale/sfpowerscripts.core/lib/sfdxwrappers/CreateUnlockedPackageImpl";
 import PackageDiffImpl from "@dxatscale/sfpowerscripts.core/lib/sfdxwrappers/PackageDiffImpl";
+import authGit from "../Common/VersionControlAuth";
 const fs = require("fs");
 import simplegit from "simple-git/promise";
 
@@ -127,77 +128,10 @@ async function run() {
 }
 
 async function pushGitTag(tagname: string): Promise<void> {
-  const version_control_provider: string = tl.getInput(
-    "versionControlProvider",
-    true
-  );
-
-  let connection: string;
-  switch (version_control_provider) {
-    case "github":
-    connection = tl.getInput("github_connection", true);
-    break;
-    case "githubEnterprise":
-    connection = tl.getInput("github_enterprise_connection", true);
-    break;
-    case "bitbucket":
-    connection = tl.getInput("bitbucket_connection", true);
-    break;
-  }
-
-  let token;
-  let username: string;
-  if (version_control_provider == "azureRepo") {
-      token = tl.getVariable("system.accessToken");
-  } else if (
-      version_control_provider == "github" ||
-      version_control_provider == "githubEnterprise"
-  ) {
-      token = tl.getEndpointAuthorizationParameter(
-      connection,
-      "AccessToken",
-      true
-      );
-  } else if (version_control_provider == "bitbucket") {
-      token = tl.getEndpointAuthorizationParameter(
-      connection,
-      "AccessToken",
-      true
-      );
-  } else {
-      username = tl.getInput("username", true);
-      token = tl.getInput("password", true);
-  }
-  //Strip https
-  const removeHttps = input => input.replace(/^https?:\/\//, "");
-
-  let repository_url = removeHttps(
-      tl.getVariable("Build.Repository.Uri")
-  );
-
-  tl.debug(`Repository URL ${repository_url}`);
-
-
   //Need to add checks to make it work in classic release pipeline
   const git = simplegit(tl.getVariable("Build.Repository.LocalPath"));
 
-  let remote: string;
-  if (version_control_provider == "azureRepo") {
-      //Fix Issue https://developercommunity.visualstudio.com/content/problem/411770/devops-git-url.html
-      repository_url = repository_url.substring(
-      repository_url.indexOf("@") + 1
-      );
-      remote = `https://x-token-auth:${token}@${repository_url}`;
-  } else if (version_control_provider == "bitbucket") {
-      remote = `https://x-token-auth:${token}@${repository_url}`;
-  } else if (
-      version_control_provider == "github" ||
-      version_control_provider == "githubEnterprise"
-  ) {
-      remote = `https://${token}:x-oauth-basic@${repository_url}`;
-  } else if (version_control_provider == "otherGit") {
-      remote = `https://${username}:${token}@${repository_url}`;
-  }
+  let remote = await authGit();
 
   await git
         .addConfig("user.name", "sfpowerscripts");
